backend/requirements.txt
fastapi
firebase-admin
python-dotenv
langchain 
transformers 
sentence-transformers 
faiss-cpu 
tree-sitter 
pygments
GitPython
tree-sitter 
tree-sitter-python 
tree-sitter-javascript
tree-sitter-languages
groq